requirements.txt
node-semver
curio
curio_http
